[naga wgsl] Move diagnostics definitions into their own module.
In `naga::common::wgsl`, move `DisplayFilterableTriggeringRule` and
diagnostics-related impls into a new sub-module, `diagonostics`.

This change is only code motion; there should be no change in
behavior.
